refactor: web 폴더에 관련 기능, 파일 이동
